requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.2
Pillow>=9.0.0

# AWS & Cloud
boto3>=1.26.0
awscli>=1.27.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.981

# Monitoring & Logging
wandb>=0.13.0
python-dotenv>=0.19.0

# Documentation
mkdocs>=1.3.0
mkdocs-material>=8.3.9